new: Added hospital directly WEST of Lochini. Added 2x transportable infantry groups at DUSHETI range. [Neck]
Intention with infantry at Dusheti is to use these for CASEVAC/MEDEVAC. For some reason I was unable to give them waypoints, so I located them at two different areas at DUSHETI to give the JTAC flexibility.

Hospital is a new mod. I will post details and NOTAM on 132nd website to make sure everyone have the hospital. The hospital have a landing pad for RW coming it with patients.